Module mint_client::api

Modules

fake

Structs

DynFederationApi

FederationError

An API request error when calling an entire federation

FederationMember 🔒

PeerResponse

ValidHistoryWrapper 🔒

WsFederationApi

Mint API client that will try to run queries against all members expecting equal results from at least min_eq_results of them. Members that return differing results are returned as a member faults list.

WsFederationConnect

Information required for client to construct WsFederationApi instance

__FieldVisitor 🔒

__Visitor 🔒

Enums

MemberError

An API request error when calling a single federation member

OutputOutcomeError

__Field 🔒

Constants

FIELDS 🔒

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

Traits

FederationApiExt

An extension trait allowing to making federation-wide API call on top IFederationApi.

GlobalFederationApi

IFederationApi

JsonRpcClient

LnFederationApi

MintFederationApi

WalletFederationApi

Functions

erased_multi_param

Build a Vec<json::Value> that IFederationApi::request_raw expects when multiple argument are passed to the API call

erased_no_param

Build a Vec<json::Value> that IFederationApi::request_raw expects when no arguments are passed to the API call

erased_single_param

Build a Vec<json::Value> that IFederationApi::request_raw expects when one argument are passed to the API call

url_to_string_with_default_port 🔒

jsonrpsee converts the Url to a &str internally and then parses it as an Uri. Unfortunately Url swallows ports that it considers default ports (e.g. 80 and 443 for HTTP(S)) which makes the Uri parsing fail in these cases. This function works around this limitation in a limited way (not fully standard compliant, but work for our use case).

Type Definitions

FederationResult

JsonRpcResult

JsonValue 🔒

MemberResult

OutputOutcomeResult 🔒

Opaque Types

🔒