Struct mint_client::api::fake::FederationApiFaker

pub struct FederationApiFaker<State> {
    state: Arc<State>,
    members: BTreeSet<PeerId>,
    handlers: BTreeMap<String, Pin<Box<dyn Fn(Arc<State>, Vec<Value>) -> Pin<Box<dyn Future<Output = RpcResult<Value>> + Send>> + Send + Sync>>>,
}

A fake super::IFederationApi builder

This struct allows easily stubbing responses to given API calls, by listing a list of handlers for methods that are expected to be.

Fields
state: Arc<State>members: BTreeSet<PeerId>handlers: BTreeMap<String, Pin<Box<dyn Fn(Arc<State>, Vec<Value>) -> Pin<Box<dyn Future<Output = RpcResult<Value>> + Send>> + Send + Sync>>>
Implementations

impl<State> FederationApiFaker<State>where
State: Debug,

pub fn new(state: Arc<State>, members: BTreeSet<PeerId>) -> Self

pub fn with<F, Fut, Param, Ret>(self, method: impl Into<String>, f: F) -> Selfwhere
State: Send + Sync + 'static,
F: Fn(Arc<State>, Param) -> Fut + Send + Sync + 'static + Copy,
Fut: Future<Output = RpcResult<Ret>> + Send + 'static,
Param: DeserializeOwned + Send + Sync,
Ret: Serialize,

Add a handler f to a method call