Struct mint_client::clients::user::UserClient

pub struct UserClient {
    context: OwnedClientContext<ClientConfig>,
}

Fields
context: OwnedClientContext<ClientConfig>
Implementations

impl UserClient

pub fn new(
cfg: ClientConfig,
db: Box<dyn Database>,
secp: Secp256k1<All>
) -> Self

pub fn new_with_api(
config: ClientConfig,
db: Box<dyn Database>,
api: Box<dyn FederationApi>,
secp: Secp256k1<All>
) -> UserClient

pub fn ln_client(&self) -> LnClient<'_>

pub fn mint_client(&self) -> MintClient<'_>

pub fn wallet_client(&self) -> WalletClient<'_>

pub async fn peg_in<R: RngCore + CryptoRng>(
&self,
txout_proof: TxOutProof,
btc_transaction: BitcoinTransaction,
rng: R
) -> Result<TransactionId, ClientError>

async fn submit_tx_with_change<R: RngCore + CryptoRng>(
&self,
tx: TransactionBuilder,
batch: Accumulator<BatchItem>,
rng: R
) -> Result<TransactionId, ClientError>

pub async fn reissue<R: RngCore + CryptoRng>(
&self,
coins: Coins<SpendableCoin>,
rng: R
) -> Result<OutPoint, ClientError>

Exchanges coins received from an untrusted third party for newly issued ones to prevent double spends. Users must ensure that the reissuance transaction is accepted before accepting coins as a valid payment.

On success the out point of the newly issued e-cash tokens is returned. It can be used to easily poll the transaction status using MintClient::fetch_coins until it returns Ok(()), indicating we received our newly issued e-cash tokens.

pub async fn pay_for_coins<R: RngCore + CryptoRng>(
&self,
coins: Coins<BlindToken>,
rng: R
) -> Result<OutPoint, ClientError>

pub fn receive_coins<R: RngCore + CryptoRng>(
&self,
amount: Amount,
rng: R,
create_tx: impl FnMut(Coins<BlindToken>) -> OutPoint
)

pub async fn peg_out<R: RngCore + CryptoRng>(
&self,
amt: Amount,
address: Address,
rng: R
) -> Result<TransactionId, ClientError>

pub fn get_new_pegin_address<R: RngCore + CryptoRng>(&self, rng: R) -> Address

pub fn select_and_spend_coins(
&self,
amount: Amount
) -> Result<Coins<SpendableCoin>, MintClientError>

pub async fn fetch_coins<'a>(
&self,
outpoint: OutPoint
) -> Result<(), MintClientError>

Tries to fetch e-cash tokens from a certain out point. An error may just mean having queried the federation too early. Use MintClientError::is_retryable to determine if the operation should be retried at a later time.

pub async fn fetch_all_coins<'a>(
&self
) -> Result<Vec<TransactionId>, MintClientError>

pub fn coins(&self) -> Coins<SpendableCoin>

pub async fn fund_outgoing_ln_contract<R: RngCore + CryptoRng>(
&self,
gateway: &LightningGateway,
invoice: Invoice,
rng: R
) -> Result<(ContractId, OutPoint), ClientError>

pub async fn await_outgoing_contract_acceptance(
&self,
outpoint: OutPoint
) -> Result<(), ClientError>

pub async fn create_unconfirmed_invoice<R: RngCore + CryptoRng>(
&self,
amount: Amount,
description: String,
rng: R
) -> Result<(KeyPair, UnconfirmedInvoice), ClientError>

pub async fn confirm_invoice(
&self,
invoice: UnconfirmedInvoice
) -> Result<Invoice, ClientError>

Wait for this invoice’s offer to be accepted by federation

pub async fn claim_incoming_contract(
&self,
contract_id: ContractId,
keypair: KeyPair,
rng: impl RngCore + CryptoRng
) -> Result<TransactionId, ClientError>