Struct mint_client::ln::LnClient

pub struct LnClient {
    pub config: LightningClientConfig,
    pub context: Arc<ClientContext>,
}

Fields
config: LightningClientConfigcontext: Arc<ClientContext>
Implementations

impl LnClient

pub async fn create_outgoing_output<'a, 'b>(
&'a self,
dbtx: &mut DatabaseTransaction<'b>,
invoice: Invoice,
gateway: &LightningGateway,
timelock: u32,
rng: impl RngCore + CryptoRng + 'a
) -> Result<LightningOutput>

Create an output that incentivizes a Lighning gateway to pay an invoice for us. It has time till the block height defined by timelock, after that we can claim our money back.

pub async fn get_contract_account(
&self,
id: ContractId
) -> Result<ContractAccount>

pub async fn get_outgoing_contract(
&self,
id: ContractId
) -> Result<OutgoingContractAccount>

pub async fn is_outgoing_contract_refundable(
&self,
id: ContractId
) -> Result<bool>

Determines if an outgoing contract can be refunded

pub async fn await_outgoing_refundable(&self, id: ContractId) -> Result<()>

Waits for an outgoing contract to become refundable

pub async fn get_incoming_contract(
&self,
id: ContractId
) -> Result<IncomingContractAccount>

pub async fn refundable_outgoing_contracts(
&self,
block_height: u64
) -> Vec<OutgoingContractData>

pub fn create_refund_outgoing_contract_input<'a>(
&self,
contract_data: &'a OutgoingContractData
) -> (&'a KeyPair, LightningInput)

pub fn create_offer_output(
&self,
amount: Amount,
payment_hash: Sha256Hash,
payment_secret: Preimage,
expiry_time: Option<u64>
) -> LightningOutput

pub async fn get_offer(
&self,
payment_hash: Sha256Hash
) -> Result<IncomingContractOffer>

pub async fn offer_exists(&self, payment_hash: Sha256Hash) -> Result<bool>

pub async fn save_confirmed_invoice(&self, invoice: &ConfirmedInvoice)

pub async fn get_confirmed_invoice(
&self,
contract_id: ContractId
) -> Result<ConfirmedInvoice>

pub fn create_cancel_outgoing_output(
&self,
contract_id: ContractId,
signature: Signature
) -> LightningOutput

Used by gateway to prematurely return funds to the user if the payment failed