Module mint_client::mint

Modules

db

Structs

MintClient

Federation module client for the Mint module. It can both create transaction inputs and outputs of the mint type.

NoteIssuanceRequest

Single Note issuance request to the mint.

NoteIssuanceRequests

Multiple Note issuance requests

SpendableNote

A Note with associated secret key that allows to proof ownership (spend it)

Enums

CoinFinalizationError

MintClientError

Type Definitions

Result 🔒