Module mint_client::mint::backup

Backup and recovery of ecash notes

Ecash backup is implemented by periodically taking a snapshot, self-encrypting it and uploading it to federation.

Recovery is done by deriving deterministic ecash nonces and checking, them with federation. A successfully recovered snapshot can be used to avoid having to scan the whole history.

Structs

EcashBackup

Encrypted version of PlaintextEcashBackup.

EcashRecoveryFinalState

EcashRecoveryTracker ğŸ”’

The state machine used for fast-fowarding backup from point when it was taken to the present time by following epoch history items from the time the snapshot was taken.

PlaintextEcashBackup

Snapshot of a ecash state (notes)

__FieldVisitor ğŸ”’

__Visitor ğŸ”’

Enums

__Field ğŸ”’

Constants

FIELDS ğŸ”’

Statics

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

Opaque Types

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’