Struct mint_client::Client

pub struct Client<C> {
    pub(crate) config: C,
    pub(crate) context: ClientContext,
}

Fields
config: Ccontext: ClientContext
Implementations

impl<T: AsRef<ClientConfig> + Clone> Client<T>

pub fn ln_client(&self) -> LnClient<'_>

pub fn mint_client(&self) -> MintClient<'_>

pub fn wallet_client(&self) -> WalletClient<'_>

pub fn config(&self) -> T

pub fn new(config: T, db: Box<dyn Database>, secp: Secp256k1<All>) -> Self

pub fn new_with_api(
config: T,
db: Box<dyn Database>,
api: Box<dyn FederationApi>,
secp: Secp256k1<All>
) -> Client<T>

pub async fn peg_in<R: RngCore + CryptoRng>(
&self,
txout_proof: TxOutProof,
btc_transaction: BitcoinTransaction,
rng: R
) -> Result<TransactionId, ClientError>

pub(crate) async fn submit_tx_with_change<R: RngCore + CryptoRng>(
&self,
tx: TransactionBuilder,
batch: Accumulator<BatchItem>,
rng: R
) -> Result<TransactionId, ClientError>

pub async fn reissue<R: RngCore + CryptoRng>(
&self,
coins: TieredMulti<SpendableCoin>,
rng: R
) -> Result<OutPoint, ClientError>

Exchanges coins received from an untrusted third party for newly issued ones to prevent double spends. Users must ensure that the reissuance transaction is accepted before accepting coins as a valid payment.

On success the out point of the newly issued e-cash tokens is returned. It can be used to easily poll the transaction status using MintClient::fetch_coins until it returns Ok(()), indicating we received our newly issued e-cash tokens.

pub async fn validate_tokens(
&self,
coins: &TieredMulti<SpendableCoin>
) -> Result<(), ClientError>

Validate tokens without claiming them. This function checks if signatures are valid based on the federation public key. It does not check if the nonce is unspent.

pub async fn pay_for_coins<R: RngCore + CryptoRng>(
&self,
coins: TieredMulti<BlindToken>,
rng: R
) -> Result<OutPoint, ClientError>

pub fn receive_coins<R: RngCore + CryptoRng>(
&self,
amount: Amount,
rng: R,
create_tx: impl FnMut(TieredMulti<BlindToken>) -> OutPoint
)

pub async fn new_peg_out_with_fees(
&self,
amount: Amount,
recipient: Address
) -> Result<PegOut, ClientError>

pub async fn peg_out<R: RngCore + CryptoRng>(
&self,
peg_out: PegOut,
rng: R
) -> Result<OutPoint, ClientError>

pub fn get_new_pegin_address<R: RngCore + CryptoRng>(&self, rng: R) -> Address

Returns a bitcoin address suited to perform a fedimint peg-in

This function requires a cryptographically secure randomness source, and utilizes the wallet-clients get_new_pegin_address to derive a bitcoin-address from the federations public descriptor by tweaking it.

this function will write to the clients DB

read more on fedimints address derivation: https://fedimint.org/Fedimint/wallet/

pub fn select_and_spend_coins(
&self,
amount: Amount
) -> Result<TieredMulti<SpendableCoin>, ClientError>

WARNING this selects and removes coins from the database without confirming whether we have successfully spent them in a transaction.

pub async fn fetch_coins<'a>(
&self,
outpoint: OutPoint
) -> Result<(), ClientError>

Tries to fetch e-cash tokens from a certain out point. An error may just mean having queried the federation too early. Use MintClientError::is_retryable to determine if the operation should be retried at a later time.

pub async fn reissue_pending_coins<R: RngCore + CryptoRng>(
&self,
rng: R
) -> Result<OutPoint, ClientError>

Should be called after any transaction that might have failed in order to get any coin inputs back.

pub async fn await_consensus_block_height(&self, block_height: u64) -> u64

pub async fn fetch_all_coins<'a>(&self) -> Vec<Result<OutPoint, ClientError>>

pub fn coins(&self) -> TieredMulti<SpendableCoin>

pub fn list_active_issuances(&self) -> Vec<(OutPoint, CoinFinalizationData)>

pub async fn fetch_epoch_history(
&self,
epoch: u64
) -> Result<EpochHistory, ClientError>

impl Client<UserClientConfig>

pub async fn fetch_registered_gateways(
&self
) -> Result<Vec<LightningGateway>, ClientError>

pub async fn fetch_active_gateway(
&self
) -> Result<LightningGateway, ClientError>

pub async fn switch_active_gateway(
&self,
node_pub_key: Option<PublicKey>
) -> Result<LightningGateway, ClientError>

Switches the clients active gateway to a registered gateway with the given node pubkey. If no pubkey is given (node_pub_key == None) the first available registered gateway is activated. This behavior is useful for scenarios where we don’t know any registered gateways in advance.

pub async fn fund_outgoing_ln_contract<R: RngCore + CryptoRng>(
&self,
invoice: Invoice,
rng: R
) -> Result<(ContractId, OutPoint), ClientError>

pub async fn await_outgoing_contract_acceptance(
&self,
outpoint: OutPoint
) -> Result<(), ClientError>

pub async fn generate_invoice<R: RngCore + CryptoRng>(
&self,
amount: Amount,
description: String,
rng: R
) -> Result<ConfirmedInvoice, ClientError>

pub async fn claim_incoming_contract(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<OutPoint, ClientError>

pub async fn await_outgoing_contract_execution(
&self,
contract_id: ContractId
) -> Result<(), ClientError>

Notify gateway that we’ve escrowed tokens they can claim by routing our payment and wait for them to do so

impl Client<GatewayClientConfig>

pub async fn fetch_outgoing_contract(
&self,
contract_id: ContractId
) -> Result<OutgoingContractAccount, ClientError>

Fetch the specified outgoing payment contract account

pub async fn validate_outgoing_account(
&self,
account: &OutgoingContractAccount
) -> Result<PaymentParameters, ClientError>

Check if we can claim the contract account and returns the max delay in blocks for how long other nodes on the route are allowed to delay the payment.

pub fn save_outgoing_payment(&self, contract: OutgoingContractAccount)

Save the details about an outgoing payment the client is about to process. This function has to be called prior to instructing the lightning node to pay the invoice since otherwise a crash could lead to loss of funds.

Note though that extended periods of staying offline will result in loss of funds anyway if the client can not claim the respective contract in time.

pub fn list_pending_outgoing(&self) -> Vec<OutgoingContractAccount>

Lists all previously saved transactions that have not been driven to completion so far

pub fn abort_outgoing_payment(&self, contract_id: ContractId)

Abort payment if our node can’t route it

pub async fn claim_outgoing_contract(
&self,
contract_id: ContractId,
preimage: [u8; 32],
rng: impl RngCore + CryptoRng
) -> Result<OutPoint, ClientError>

Claim an outgoing contract after acquiring the preimage by paying the associated invoice and initiates e-cash issuances to receive the bitcoin from the contract (these still need to be fetched later to finalize them).

Callers need to make sure that the contract can still be claimed by the gateway and has not timed out yet. Otherwise the transaction will fail.

pub async fn buy_preimage_offer(
&self,
payment_hash: &Hash,
amount: &Amount,
rng: impl RngCore + CryptoRng
) -> Result<(OutPoint, ContractId), ClientError>

pub async fn refund_incoming_contract(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<TransactionId, ClientError>

Claw back funds after outgoing contract that had invalid preimage

pub fn list_pending_claimed_outgoing(&self) -> Vec<ContractId>

Lists all claim transactions for outgoing contracts that we have submitted but were not part of the consensus yet.

pub async fn await_preimage_decryption(
&self,
outpoint: OutPoint
) -> Result<Preimage, ClientError>

Wait for a lightning preimage gateway has purchased to be decrypted by the federation

pub async fn await_outgoing_contract_claimed(
&self,
contract_id: ContractId,
outpoint: OutPoint
) -> Result<(), ClientError>

Waits for a outgoing contract claim transaction to be confirmed and retransmits it periodically if this does not happen.

pub fn list_fetchable_coins(&self) -> Vec<OutPoint>

pub async fn register_with_federation(
&self,
config: LightningGateway
) -> Result<(), ClientError>

Register this gateway with the federation