Struct mint_client::mint::MintClient

pub struct MintClient {
    pub epoch_pk: PublicKey,
    pub config: MintClientConfig,
    pub context: Arc<ClientContext>,
    pub secret: DerivableSecret,
}

Federation module client for the Mint module. It can both create transaction inputs and outputs of the mint type.

Fields
epoch_pk: PublicKeyconfig: MintClientConfigcontext: Arc<ClientContext>secret: DerivableSecret
Implementations

impl MintClient

pub async fn back_up_ecash_to_federation(&self) -> Result<()>

Prepare an encrypted backup and send it to federation for storing

pub async fn download_ecash_backup_from_federation(
&self
) -> Result<PlaintextEcashBackup>

fn get_derived_backup_encryption_key_static(
secret: &DerivableSecret
) -> LessSafeKey

Static version of Self::get_derived_backup_encryption_key for testing without creating whole MintClient

fn get_derived_backup_signing_key_static(secret: &DerivableSecret) -> KeyPair

Static version of Self::get_derived_backup_signing_key for testing without creating whole MintClient

fn get_derived_backup_encryption_key(&self) -> LessSafeKey

fn get_derived_backup_signing_key(&self) -> KeyPair

async fn prepare_plaintext_ecash_backup(&self) -> Result<PlaintextEcashBackup>

async fn prepare_ecash_backup(&self) -> Result<EcashBackup>

async fn upload_ecash_backup(&self, backup: EcashBackup) -> Result<()>

async fn fetch_epochs(
&self,
epoch_range: RangeInclusive<u64>,
sender: Sender<Result<EpochHistory>>,
task_handle: &TaskHandle
)

Fetch epochs in a given range and send them over sender

Since WASM’s spawn does not support join handles, we indicate errors via sender itself.

TODO: could be internal to recovery_loop?

pub async fn recovery_loop(
&self,
task_group: &mut TaskGroup,
backup: PlaintextEcashBackup,
pub_key_shares: BTreeMap<PeerId, Tiered<PublicKeyShare>>
) -> Cancellable<Result<()>>

impl MintClient

pub async fn start_dbtx(&self) -> DatabaseTransaction<'_>

pub async fn finalize_change(&self, tx: &mut Transaction, change: Vec<Amount>)

Adds the final amounts of change to the tx before submitting it Allows for multiple change outputs

async fn create_ecash(
&self,
amount: Amount
) -> (NoteIssuanceRequests, TieredMulti<BlindNonce>)

Generates unsigned ecash, along with the private keys that can spend it

pub async fn select_input(
&self,
amount: Amount
) -> Result<(Vec<KeyPair>, Input), MintClientError>

pub fn ecash_input(
ecash: TieredMulti<SpendableNote>
) -> Result<(Vec<KeyPair>, Input), MintClientError>

pub async fn coins(&self) -> TieredMulti<SpendableNote>

pub async fn get_available_notes(
&self,
dbtx: &mut DatabaseTransaction<'_>
) -> TieredMulti<SpendableNote>

Get available spendable notes with a db transaction already opened

pub async fn get_next_note_index(
&self,
dbtx: &mut DatabaseTransaction<'_>,
amount: Amount
) -> NoteIndex

async fn new_note_secret(&self, amount: Amount) -> DerivableSecret

pub fn new_note_secret_static(
secret: &DerivableSecret,
amount: Amount,
note_idx: NoteIndex
) -> DerivableSecret

Derive the note DerivableSecret from the Mint’s secret the amount tier and note_idx

Static to help re-use in other places, that don’t have a whole Self available

pub async fn new_ecash_note<C: Signing>(
&self,
ctx: &Secp256k1<C>,
amount: Amount
) -> (NoteIssuanceRequest, BlindNonce)

pub async fn select_coins(
&self,
amount: Amount
) -> Result<TieredMulti<SpendableNote>, MintClientError>

pub async fn receive_coins<'a, F, Fut>(
&self,
amount: Amount,
dbtx: &mut DatabaseTransaction<'a>,
create_tx: F
)where
F: FnMut(TieredMulti<BlindNonce>) -> Fut,
Fut: Future<Output = OutPoint>,

pub async fn fetch_coins<'a>(
&self,
dbtx: &mut DatabaseTransaction<'a>,
outpoint: OutPoint
) -> Result<(), MintClientError>

pub async fn list_active_issuances(
&self
) -> Vec<(OutPoint, NoteIssuanceRequests)>

pub async fn fetch_all_coins(&self) -> Vec<Result<OutPoint, MintClientError>>