Struct mint_client::mint::MintClient

pub struct MintClient<'c> {
    pub config: MintClientConfig,
    pub context: &'c ClientContext,
    pub secret: DerivableSecret,
}

Federation module client for the Mint module. It can both create transaction inputs and outputs of the mint type.

Fields
config: MintClientConfigcontext: &'c ClientContextsecret: DerivableSecret
Implementations

impl<'c> MintClient<'c>

pub async fn back_up_ecash_to_federation(&self) -> Result<()>

Prepare an encrypted backup and send it to federation for storing

pub async fn download_ecash_backup_from_federation(
&self
) -> Result<PlaintextEcashBackup>

fn get_derived_backup_encryption_key_static(
secret: &DerivableSecret
) -> LessSafeKey

Static version of Self::get_derived_backup_encryption_key for testing without creating whole MintClient

fn get_derived_backup_signing_key_static(secret: &DerivableSecret) -> KeyPair

Static version of Self::get_derived_backup_signing_key for testing without creating whole MintClient

fn get_derived_backup_encryption_key(&self) -> LessSafeKey

fn get_derived_backup_signing_key(&self) -> KeyPair

async fn prepare_plaintext_ecash_backup(&self) -> Result<PlaintextEcashBackup>

async fn prepare_ecash_backup(&self) -> Result<EcashBackup>

async fn upload_ecash_backup(&self, backup: EcashBackup) -> Result<()>

impl<'c> MintClient<'c>

pub fn start_dbtx(&self) -> DatabaseTransaction<'_>

pub fn coins(&self) -> TieredMulti<SpendableNote>

pub fn get_available_notes(
&self,
dbtx: &mut DatabaseTransaction<'_>
) -> TieredMulti<SpendableNote>

Get available spendable notes with a db transaction already opened

pub fn get_last_note_index(
&self,
dbtx: &mut DatabaseTransaction<'_>
) -> NoteIndex

fn new_note_secret(&self, dbtx: &mut DatabaseTransaction<'_>) -> DerivableSecret

pub fn new_ecash_note<C: Signing>(
&self,
ctx: &Secp256k1<C>,
dbtx: &mut DatabaseTransaction<'_>
) -> (NoteIssuanceRequest, BlindNonce)

pub fn select_coins(
&self,
amount: Amount
) -> Result<TieredMulti<SpendableNote>, MintClientError>

pub async fn submit_tx_with_change<'a, C, R>(
&self,
client: &Client<C>,
tx: TransactionBuilder,
rng: R
) -> Result<TransactionId, MintClientError>where
C: AsRef<ClientConfig> + Clone,
R: RngCore + CryptoRng,

pub async fn receive_coins<'a, F, Fut>(
&self,
amount: Amount,
dbtx: &mut DatabaseTransaction<'a>,
coin_gen: impl Fn(&mut DatabaseTransaction<'_>) -> (NoteIssuanceRequest, BlindNonce),
create_tx: F
)where
F: FnMut(TieredMulti<BlindNonce>) -> Fut,
Fut: Future<Output = OutPoint>,

pub async fn fetch_coins<'a>(
&self,
dbtx: &mut DatabaseTransaction<'a>,
outpoint: OutPoint
) -> Result<(), MintClientError>

pub fn list_active_issuances(&self) -> Vec<(OutPoint, NoteIssuanceRequests)>

pub async fn fetch_all_coins(&self) -> Vec<Result<OutPoint, MintClientError>>