Struct mint_client::Client

pub struct Client<C> {
    pub(crate) config: C,
    pub(crate) context: ClientContext,
    pub(crate) root_secret: DerivableSecret,
}

Fields
config: Ccontext: ClientContextroot_secret: DerivableSecret
Implementations

impl<T: AsRef<ClientConfig> + Clone> Client<T>

pub fn ln_client(&self) -> LnClient<'_>

pub fn mint_client(&self) -> MintClient<'_>

pub fn wallet_client(&self) -> WalletClient<'_>

pub fn config(&self) -> T

pub fn new(config: T, db: Database, secp: Secp256k1<All>) -> Self

pub fn new_with_api(
config: T,
db: Database,
api: FederationApi,
secp: Secp256k1<All>
) -> Client<T>

pub(crate) fn get_secret(db: &Database) -> DerivableSecret

Fetches the client secret from the database or generates a new one if none is present

pub async fn peg_in<R: RngCore + CryptoRng>(
&self,
txout_proof: TxOutProof,
btc_transaction: BitcoinTransaction,
rng: R
) -> Result<TransactionId, ClientError>

pub(crate) async fn submit_tx_with_change<R: RngCore + CryptoRng>(
&self,
tx: TransactionBuilder,
rng: R
) -> Result<TransactionId, ClientError>

pub async fn reissue<R: RngCore + CryptoRng>(
&self,
notes: TieredMulti<SpendableNote>,
rng: R
) -> Result<OutPoint, ClientError>

Spent some SpendableNotes to receive a freshly minted ones

This is useful in scenarios where certain notes were handed over directly to us by another user as a payment. By spending them we can make sure they can no longer be potentially double-spent.

On success the out point of the newly issued e-cash tokens is returned. It can be used to easily poll the transaction status using MintClient::fetch_coins until it returns Ok(()), indicating we received our newly issued e-cash tokens.

pub async fn validate_note_signatures(
&self,
notes: &TieredMulti<SpendableNote>
) -> Result<(), ClientError>

Validate signatures on notes.

This function checks if signatures are valid based on the federation public key. It does not check if the nonce is unspent.

pub async fn pay_to_blind_nonces<R: RngCore + CryptoRng>(
&self,
blind_nonces: TieredMulti<BlindNonce>,
rng: R
) -> Result<OutPoint, ClientError>

Pay by creating notes provided (and most probably controlled) by the recipient.

A standard way to facilitate a payment between users of a mint. Generate a transaction spending notes we own as inputs and creating new notes from [BlindNonce]s provided by the recipient as outputs.

Returns a OutPoint of a fedimint transaction created and submitted as a payment.

The name is derived from Bitcoin’s terminology of “pay to <address-type>”.

pub async fn receive_coins<F, Fut>(&self, amount: Amount, create_tx: F)where
F: FnMut(TieredMulti<BlindNonce>) -> Fut,
Fut: Future<Output = OutPoint>,

pub async fn new_peg_out_with_fees(
&self,
amount: Amount,
recipient: Address
) -> Result<PegOut, ClientError>

pub async fn peg_out<R: RngCore + CryptoRng>(
&self,
peg_out: PegOut,
rng: R
) -> Result<OutPoint, ClientError>

pub async fn get_new_pegin_address<R: RngCore + CryptoRng>(
&self,
rng: R
) -> Address

Returns a bitcoin address suited to perform a fedimint peg-in

This function requires a cryptographically secure randomness source, and utilizes the wallet-clients get_new_pegin_address to derive a bitcoin-address from the federations public descriptor by tweaking it.

this function will write to the clients DB

read more on fedimints address derivation: https://fedimint.org/Fedimint/wallet/

pub async fn spend_ecash<R: RngCore + CryptoRng>(
&self,
amount: Amount,
rng: R
) -> Result<TieredMulti<SpendableNote>, ClientError>

Issues a spendable amount of ecash

WARNING the ecash will be deleted from the database, the returned ecash must be reissued or it will be lost

pub async fn fetch_coins<'a>(
&self,
outpoint: OutPoint
) -> Result<(), ClientError>

Tries to fetch e-cash tokens from a certain out point. An error may just mean having queried the federation too early. Use MintClientError::is_retryable to determine if the operation should be retried at a later time.

pub async fn reissue_pending_coins<R: RngCore + CryptoRng>(
&self,
rng: R
) -> Result<OutPoint, ClientError>

Should be called after any transaction that might have failed in order to get any coin inputs back.

pub async fn await_consensus_block_height(
&self,
block_height: u64
) -> Result<u64, Elapsed>

pub(crate) async fn await_consensus_block_height_inner(
&self,
block_height: u64
) -> u64

pub async fn fetch_all_coins<'a>(&self) -> Vec<Result<OutPoint, ClientError>>

pub fn coins(&self) -> TieredMulti<SpendableNote>

pub fn list_active_issuances(&self) -> Vec<(OutPoint, NoteIssuanceRequests)>

pub async fn fetch_epoch_history(
&self,
epoch: u64,
epoch_pk: PublicKey
) -> Result<EpochHistory, ClientError>

impl Client<UserClientConfig>

pub async fn fetch_registered_gateways(
&self
) -> Result<Vec<LightningGateway>, ClientError>

pub async fn fetch_active_gateway(
&self
) -> Result<LightningGateway, ClientError>

pub async fn switch_active_gateway(
&self,
node_pub_key: Option<PublicKey>
) -> Result<LightningGateway, ClientError>

Switches the clients active gateway to a registered gateway with the given node pubkey. If no pubkey is given (node_pub_key == None) the first available registered gateway is activated. This behavior is useful for scenarios where we don’t know any registered gateways in advance.

pub async fn fund_outgoing_ln_contract<R: RngCore + CryptoRng>(
&self,
invoice: Invoice,
rng: R
) -> Result<(ContractId, OutPoint), ClientError>

pub async fn try_refund_outgoing_contract(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<OutPoint, ClientError>

Claims a refund for an expired or cancelled outgoing contract

This can be necessary when the Lightning gateway cannot route the payment, is malicious or offline. The function returns the out point of the e-cash output generated as change.

pub async fn await_outgoing_contract_acceptance(
&self,
outpoint: OutPoint
) -> Result<(), ClientError>

pub async fn await_outpoint_outcome(
&self,
outpoint: OutPoint
) -> Result<(), ClientError>

Waits for the federation to sign an ecash note.

This function will poll until the returned result includes a SigResponse from the federation or it will timeout.

pub async fn generate_invoice<R: RngCore + CryptoRng>(
&self,
amount: Amount,
description: String,
rng: R,
expiry_time: Option<u64>
) -> Result<ConfirmedInvoice, ClientError>

pub async fn claim_incoming_contract(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<OutPoint, ClientError>

pub async fn await_outgoing_contract_execution(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<(), ClientError>

Notify gateway that we’ve escrowed tokens they can claim by routing our payment and wait for them to do so

impl Client<GatewayClientConfig>

pub async fn fetch_outgoing_contract(
&self,
contract_id: ContractId
) -> Result<OutgoingContractAccount, ClientError>

Fetch the specified outgoing payment contract account

pub async fn validate_outgoing_account(
&self,
account: &OutgoingContractAccount
) -> Result<PaymentParameters, ClientError>

Check if we can claim the contract account and returns the max delay in blocks for how long other nodes on the route are allowed to delay the payment.

pub(crate) fn is_maybe_internal_payment(&self, invoice: &Invoice) -> bool

Returns true if the invoice contains us as a routing hint

pub async fn save_outgoing_payment(&self, contract: OutgoingContractAccount)

Save the details about an outgoing payment the client is about to process. This function has to be called prior to instructing the lightning node to pay the invoice since otherwise a crash could lead to loss of funds.

Note though that extended periods of staying offline will result in loss of funds anyway if the client can not claim the respective contract in time.

pub fn list_pending_outgoing(&self) -> Vec<OutgoingContractAccount>

Lists all previously saved transactions that have not been driven to completion so far

pub async fn abort_outgoing_payment(
&self,
contract_id: ContractId
) -> Result<(), ClientError>

Abort payment if our node can’t route it and give money back to user

pub async fn claim_outgoing_contract(
&self,
contract_id: ContractId,
preimage: Preimage,
rng: impl RngCore + CryptoRng
) -> Result<OutPoint, ClientError>

Claim an outgoing contract after acquiring the preimage by paying the associated invoice and initiates e-cash issuances to receive the bitcoin from the contract (these still need to be fetched later to finalize them).

Callers need to make sure that the contract can still be claimed by the gateway and has not timed out yet. Otherwise the transaction will fail.

pub async fn buy_preimage_offer(
&self,
payment_hash: &Hash,
htlc_amount: &Amount,
rng: impl RngCore + CryptoRng
) -> Result<(OutPoint, ContractId), ClientError>

Buy a lightning preimage listed for sale inside the federation

Called when a lightning gateway attempts to satisfy a contract on behalf of a user

payment_hash - hash of the preimage we want to buy. It is included inside a bolt11 invoice and should match the offer hash

htlc_amount - amount from the htlc the gateway wants to pay. Should be less than or equal to the offer amount depending on gateway fee policy

pub async fn refund_incoming_contract(
&self,
contract_id: ContractId,
rng: impl RngCore + CryptoRng
) -> Result<TransactionId, ClientError>

Claw back funds after incoming contract that had invalid preimage

pub fn list_pending_claimed_outgoing(&self) -> Vec<ContractId>

Lists all claim transactions for outgoing contracts that we have submitted but were not part of the consensus yet.

pub async fn await_preimage_decryption(
&self,
outpoint: OutPoint
) -> Result<Preimage, ClientError>

Wait for a lightning preimage gateway has purchased to be decrypted by the federation

pub async fn await_outgoing_contract_claimed(
&self,
contract_id: ContractId,
outpoint: OutPoint
) -> Result<(), ClientError>

Waits for a outgoing contract claim transaction to be confirmed and retransmits it periodically if this does not happen.

pub fn list_fetchable_coins(&self) -> Vec<OutPoint>

pub async fn register_with_federation(
&self,
config: LightningGateway
) -> Result<(), ClientError>

Register this gateway with the federation