Trait mint_client::ModuleClient

pub trait ModuleClient {
    type Module: FederationModule;

    fn input_amount(
        &self,
        input: &<Self::Module as FederationModule>::TxInput
    ) -> TransactionItemAmount;
    fn output_amount(
        &self,
        output: &<Self::Module as FederationModule>::TxOutput
    ) -> TransactionItemAmount;
}

Required Associated Types

type Module: FederationModule

Required Methods

fn input_amount(
&self,
input: &<Self::Module as FederationModule>::TxInput
) -> TransactionItemAmount

Returns the amount represented by the input and the fee its processing requires

fn output_amount(
&self,
output: &<Self::Module as FederationModule>::TxOutput
) -> TransactionItemAmount

Returns the amount represented by the output and the fee its processing requires